+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "hw2",
+  "name": "CoolSlack",
   "version": "1.0.0",
   "description": "",
   "scripts": {
