+++ public/components/Ajax/AjaxModule.js
@@ -1,0 +1,21 @@
+export const AjaxCreate = (function(){
+    class AjaxModule {
+
+        async _ajax(method, url, body = null, callback) {
+
+            let response = await fetch(url);
+            let text = await response.text();
+
+            callback(response.status, text);
+        }
+    }
+    return{
+        init:  function () {
+            globalThis.AjaxModule = new AjaxModule();
+        }
+
+    }
+
+}());
+
+
