+++ public/components/MainPage/mainPage.css
@@ -1,0 +1,36 @@
+.text{
+    padding: 50px;
+    position: fixed; top: 50%; left: 50%;
+    -webkit-transform: translate(-50%, -50%);
+    -ms-transform: translate(-50%, -50%);
+    transform: translate(-50%, -50%);
+    font-family: "Roboto", sans-serif;
+    text-transform: uppercase;
+}
+
+.stage {
+    display: flex;
+    height: 230px;
+    width: 100%;
+}
+.box {
+    align-self: flex-end;
+    animation-duration: 2s;
+    animation-iteration-count: infinite;
+    height: 20px;
+    margin: 0 auto 0 auto;
+    transform-origin: bottom;
+    width: 20px;
+
+}
+
+.bounce-2 {
+    animation: bounce-2 30s linear infinite;
+}
+@keyframes bounce-2 {
+    from { transform:rotate(0deg); }
+    to { transform:rotate(-360deg); }
+
+}
+
+
