+++ public/index.html
@@ -3,11 +3,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" href="main.css"/>
-    <link rel="stylesheet" href="slack.css"/>
     <title>CoolSlack</title>
 </head>
 <body>
 <div id="application"></div>
-<script src="main.js"></script>
+<script type="module" src="components/Ajax/AjaxModule.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
